Refactor init_app function to improve folder initialization logic and logging

+++ src/aida/app.py
@@ -30,34 +30,30 @@ class HealthCheck(BaseModel):
 
 def init_app():
     """Initialize application by ensuring necessary folders exist."""
-    logger.info("Initializing application folders")
-    # if public folder does not exist copy from src/aida/static
-    if not os.path.exists("public"):
+    if not os.path.exists("public") or not os.path.exists("config") or not os.path.exists("prompts"):
         import shutil
+        logger.info("Initializing application folders")
+        # if public folder does not exist copy from src/aida/static
+        if not os.path.exists("public"):
+            logger.info("Creating public folder from static assets")
+            # get package root
+            shutil.copytree(
+                os.path.join(os.path.dirname(__file__), "static/public"), "public"
+            )
+        # if config folder does not exist copy from src/aida/config
+        if not os.path.exists("config"):
+            logger.info("Creating config folder from static assets")
+            # get package root
+            shutil.copytree(
+                os.path.join(os.path.dirname(__file__), "static/config"), "config"
+            )
 
-        logger.info("Creating public folder from static assets")
-        # get package root
-        shutil.copytree(
-            os.path.join(os.path.dirname(__file__), "static/public"), "public"
-        )
-    # if config folder does not exist copy from src/aida/config
-    if not os.path.exists("config"):
-        import shutil
-
-        logger.info("Creating config folder from static assets")
-        # get package root
-        shutil.copytree(
-            os.path.join(os.path.dirname(__file__), "static/config"), "config"
-        )
-
-    if not os.path.exists("prompts"):
-        import shutil
-
-        logger.info("Creating prompts folder from static assets")
-        # get package root
-        shutil.copytree(
-            os.path.join(os.path.dirname(__file__), "static/prompts"), "prompts"
-        )
+        if not os.path.exists("prompts"):
+            logger.info("Creating prompts folder from static assets")
+            # get package root
+            shutil.copytree(
+                os.path.join(os.path.dirname(__file__), "static/prompts"), "prompts"
+            )
 
 
 def create_app() -> FastAPI:
